Clarify :packages argument
The argument to :packages is the same as the FILES argument to eval-after-load.
Clarify that it can be a list of files or features and add a :files alias if
anyone would rather the syntax match eval-after-load.  Keep the packages
terminology though as I think it's more intuitive (and short compared to
something like :files-or-features).